src/Tripletex_AI_Accounting_Agent/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
httpx>=0.25.0
google-genai>=1.0.0
pydantic>=2.0.0
pdfplumber>=0.10.0
